refactor(backend): Move passkey routes to their own file
