+++ backend/app/main.py
@@ -10,7 +10,7 @@ from tortoise.contrib.fastapi import register_tortoise
 
 from .constants import VERSION
 
-from .routes import main, auth, organizations, seasons, gamepieces, fields, match_scouting, pit_scouting, data, event, uploads, emails, reports, profile, repairs
+from .routes import main, auth, organizations, seasons, gamepieces, fields, match_scouting, pit_scouting, data, event, uploads, emails, reports, profile, repairs, passkeys
 
 # Setup
 app: FastAPI = FastAPI(root_path="", version=VERSION)
@@ -48,6 +48,7 @@ app.include_router(emails.router)
 app.include_router(reports.router)
 app.include_router(profile.router)
 app.include_router(repairs.router)
+app.include_router(passkeys.router)
 
 # Setup database
 register_tortoise(
