+++ backend/app/schemas/passkeys.py
@@ -1,0 +1,10 @@
+from datetime import datetime
+from uuid import UUID
+
+from pydantic import BaseModel
+
+
+class PasskeyResponse(BaseModel):
+    uuid: UUID
+    label: str | None
+    created_at: datetime
